requirements.txt
requests
schedule
pymongo
python-dotenv
slacker
dnspython